Original change: https://android-review.googlesource.com/c/platform/art/+/3005156

+++ runtime/runtime_image.cc
@@ -164,7 +164,7 @@ class RuntimeImageHelper {
         heap->GetBootImagesSize(),
         boot_image_components,
         boot_image_checksums,
-        static_cast<uint32_t>(kRuntimePointerSize));
+        kRuntimePointerSize);
 
     // Data size includes everything except the bitmap and the header.
     header_.data_size_ = sections_end - sizeof(ImageHeader);
